sperrorest_parallel_commit
The recently vectorized sperrorest-function was modified to run in
parallel using the parallel package's mclapply() function. It was
renamed 'parsperrorest' and put into its own R-file. The
non-parallelized sperrorest-function was turned back to its original
state using nested for-loops, however, it could be replaced by the
lapply()-version in the future.
parsperrorest() includes the additional argument 'mc.cores', which sets
the number of parallel processing units used to run mclapply(). Its
default value is the total number of cores available on the system.