parsperrorest_RNG_benchmark_commit
RNG setup was included to achieve reproducibility. Results are
reproducible provided that par.args are the same.
Furthermore, benchmarking was included in the parsperrorest function. If
function parameter "benchmark" is TRUE, parsperrorest collects
information on execution duration as well as system information and puts
them into the output as a list() called "benchmarks" which also contains
par.args to allow for performance and efficiency evaluation.
Additionally, "library(parallel)" is now executed right after performing
some checks at the beginning of the function.